refactor: more appropriate class and folder names
